Extract `toc/collection_meta_ops.rs`
